sdlc(impl): extract hitl-handler.ts and post-pipeline.ts from engine.ts
- Files changed:
  - engine/hitl-handler.ts (new, 160 LOC): unified HITL handler for resume/detect modes
  - engine/post-pipeline.ts (new, 131 LOC): post-pipeline executor + collectPostPipelineNodes + sortPostPipelineNodes + runFailureHook
  - engine/engine.ts (refactored, 637 LOC from 849): delegates to new modules; Engine class unchanged in public API
  - engine/agent.ts: added resolveInputArtifacts (moved from engine.ts)
  - engine/config.ts: added findNodeConfig + collectAllNodeIds (moved from engine.ts)
  - engine/engine_test.ts: updated imports to new module paths
- Tests: engine/engine_test.ts updated (all 502 tests pass)
- Check: PASS
